requirements.txt
transformers
faiss-cpu
torch
gradio
requests
numpy